+++ packages/microsoft-security-adapter/test/filters/entra/app_role.test.ts
